feat(registeruser.vue): the component was integrated to VIA CEP API
The form was implemented with the VIA CEP API, to fill the form when user is searching for its zip
code. Added a button to make this search. Also some treatment on the form inputs to disable and
enable when it's needed, actually when 'street', 'neighborhood' and 'complement' doesn't have any
value.